The most significant changes involve the `CheckoutController.cs` and the addition of two new files, `EmailSender.cs` and `OrderCleanupService.cs`.
In `CheckoutController.cs`, the code has been updated to retrieve the user's ID and email from the JWT token earlier in the process. This change allows the user's email to be included when retrieving an order from the database. Additionally, a confirmation email is now sent to the user when an order is updated to `isPagato = true`.

The new `EmailSender.cs` file contains a class that implements the `IEmailSender` interface and includes a method for sending emails. This class is used in `CheckoutController.cs` to send the confirmation email.

The new `OrderCleanupService.cs` file contains a class that extends the `BackgroundService` class and includes a method for deleting all unpaid orders that are older than 15 minutes every 15 minutes. This service is added in the `Program.cs` file.

Changes:

1. `Back-End-Capstone.csproj` file updated to include a new package reference for `MailKit` version `4.4.0`.
2. `CheckoutController.cs` updated to include `Microsoft.EntityFrameworkCore` namespace and to retrieve user's ID and email from JWT token earlier.
3. User's ID retrieval code removed from later part of `CheckoutController.cs`.
4. `CheckoutController.cs` updated to include user's email when retrieving an order and to send a confirmation email when an order is updated to `isPagato = true`.
5. `Program.cs` updated to include `System.Text` namespace at the beginning and to add a hosted service for `OrderCleanupService`.
6. New `EmailSender.cs` file added with `EmailSender` class that implements `IEmailSender` interface.
7. New `OrderCleanupService.cs` file added with `OrderCleanupService` class that extends `BackgroundService` class.